azure_functions/requirements.txt
azure-functions==1.12.0
joblib==1.2.0
numpy==1.23.4
scikit-learn==1.1.3
scipy==1.9.3
threadpoolctl==3.1.0
